scripts/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
scikit-learn>=1.0.0
matplotlib>=3.0.0
numpy>=1.0.0 